requirements.txt
fastapi==0.115.12
aiocache==0.12.3
sqlalchemy==2.0.40
alembic==1.15.2
httpx==0.28.1
uvicorn[standard]
asyncpg==0.30.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-asyncio==0.26.0
openpyxl==3.1.5
starlette==0.46.2
python-multipart
typer
python-dotenv
pydantic-settings